+++ workers/unity/Packages/io.improbable.gdk.core/Subscriptions/CallbackManagers/ComponentEventCallbackManager.cs
@@ -6,10 +6,8 @@ namespace Improbable.Gdk.Subscriptions
 {
     internal class ComponentEventCallbackManager<T> : ICallbackManager where T : struct, IEvent
     {
-        private readonly IndexedCallbacks<T> callbacks = new IndexedCallbacks<T>();
+        private readonly EntityCallbacks<T> callbacks = new EntityCallbacks<T>();
         private readonly ComponentUpdateSystem componentUpdateSystem;
-
-        private ulong nextCallbackId = 1;
 
         public ComponentEventCallbackManager(World world)
         {
@@ -22,14 +20,13 @@ namespace Improbable.Gdk.Subscriptions
             for (var i = 0; i < updates.Count; ++i)
             {
                 ref readonly var update = ref updates[i];
-                callbacks.InvokeAll(update.EntityId.Id, in update.Event);
+                callbacks.InvokeAll(update.EntityId, in update.Event);
             }
         }
 
         public ulong RegisterCallback(EntityId entityId, Action<T> callback)
         {
-            callbacks.Add(entityId.Id, nextCallbackId, callback);
-            return nextCallbackId++;
+            return callbacks.Add(entityId, callback);
         }
 
         public bool UnregisterCallback(ulong callbackKey)