+++ CHANGELOG.md
@@ -12,7 +12,6 @@
 
 - GDK Tools Configuration window now autosaves. [#1356](https://github.com/spatialos/gdk-for-unity/pull/1356)
 
-
 ### Fixed
 
 - Invalid JSON is now logged if there is an error parsing the codegen output. [#1353](https://github.com/spatialos/gdk-for-unity/pull/1353)
@@ -22,6 +21,8 @@
 ### Internal
 
 - Produce code coverage reports in tests [#1359](https://github.com/spatialos/gdk-for-unity/pull/1359)
+- Replaced code generated ReferenceProviders with generic version. [#1358](https://github.com/spatialos/gdk-for-unity/pull/1358)
+- Refactor callbacks API. [#1348](https://github.com/spatialos/gdk-for-unity/pull/1348)
 
 ## `0.3.5` - 2020-04-23